Connected gloves/gants_connectes/requirements.txt
Adafruit-Blinka==7.3.0
adafruit-circuitpython-busdevice==5.1.8
adafruit-circuitpython-register==1.9.8
adafruit-circuitpython-typing==1.7.0
Adafruit-PlatformDetect==3.23.0
Adafruit-PureIO==1.1.9
bleak==0.14.3
certifi==2020.6.20
chardet==4.0.0
click==8.1.3
colorzero==1.1
configparser==5.2.0
dbus-next==0.2.3
distro==1.5.0
gpiozero==1.6.2
gTTS==2.2.4
icm20948==0.0.2
idna==2.10
joblib==1.1.0
numpy==1.22.4
pyftdi==0.54.0
pyserial==3.5
python-apt==2.2.1
pyusb==1.2.1
requests==2.25.1
rpi-ws281x==4.3.4
RPi.GPIO==0.7.0
scikit-learn==1.1.1
SciPy==1.8.1
six==1.16.0
sklearn==0.0
smbus==1.1.post2
smbus2==0.4.1
sparkfun-qwiic-i2c==0.9.11
sparkfun-qwiic-icm20948==0.0.1
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
threadpoolctl==3.1.0
urllib3==1.26.5
